We are receiving the request, but not yet uploading the document to the DB...
